python source code
In the given file the first file of learning0.py are the simple thing that says the basic things like hello world. In learning1.py there is the input of "what's your name" and you type your name and it says your name and says hello but your name come below the hello. In learning2.py it a same like learning1.py only thing is that, there is different between the hello word come along with name of yours.